requirements.txt
aiosignal==1.3.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
dm-tree==0.1.8
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
gymnasium==0.28.1
idna==3.6
imageio==2.33.1
jax-jumpy==1.0.0
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.25.2
pyarrow==14.0.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.9.0
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
svgwrite==1.4.3
sympy==1.12
tensorboardX==2.6.2.2
tifffile==2023.12.9
torch==2.1.2
Tree==0.2.4
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
